Added a type_name() function to util.cc/h
In the merge with DAP4, type_name() split into D2type_name() and
D4type_name(). That removes the symbol 'type_name()' from the library.
I added a function with that name that first tries to apply D4... and
then D2. If both fail, an exception is thrown. This provides forward
looking backward compatibility.