Changes to build and to the Test* classes.
Now configure tests for the uuid/uuid.h header (before it tested only
for the library which meant that installing the libuuid package and
not the needed libuuid-devel package would pass configure but fail
during the build). I also removed some unused code from the Test* classes.
The TestCommon.cc file was not needed.