Grammar fixes - for flex and its C++ scanner code.
There is a conflict between the defined return type of
FlexLexer::LexerInput. Some versions of flex define that
method as returning an int and some a returning a size_t.
Including generated code using one will break when a build
machine uses the other. I used dist-hook to just 'rm' those
generated lex.*.cc files  from d4_ce and d4_function - the
other generated scanners are OK. This is really a work-around
since we should be able to include the generated scanners
and then drop the build requirement for both bison and flex.