Hacked the marshT unit-test
I modified this so that it no longer relies on data being present after
serialize() has been called. It seems that none of the handlers use
this 'feature' and we can change the code so that memory is freed right
after the bits are sent fir each variable. This means that only one
variabe's data are in memory at any given time.