Added BaseType::clear_local_data()
The clear_local_data() method is used to free just the parts of a
BaseType (child of, really) that hold large amounts of data. It applies
to only a handful of the variable types (Vector, Sequence, D4Opaque and
D4Sequence). It is used to free memory right after data values are
sent to the client using serailize(). This means that data are no longer
held by the variable objects once serialize() returns - at least for
Array, Sequence, ... Other types like Byte, Int16, ..., will retain
their values because those values are not held in dynamically allocated
storage.