Added a serialize_no_release() method
This rounds out the serialize_n_release branch. The new behavior of
freeing the dynamic memory is very desireable, so it's the default,
but if some code wants to serialize, do something else and then delete
the storage, the new serialize_no_release provides that option. All
of our libdap and bes/module tests pass. This will be merged to the
master branch once the latter has been tagged.