Fix for clear_local_data() race condition
Took advantage of the current XDRStreamMarshaller impl that copies
data before writing it out. This means our code can delete the original
data held by the object right after the thread used to send that data
hass been called (actually, it could delete it _before_ starting the
thread, but doing so afterward keeps that feature out of the marshaller
class, which may come in handly later on...). Later on we may switch
to reference counted pointers and then this issue will go away.