Patch/fix/hack for XDRStreamMarshaller
This class' header requires pthread.h but that makes for issues since
we cannot put an include of config.h in a .h file that is installed.
To get on with testing, I'm checking in this version that just includes
it; later on I may modify the class to encapsulate the pthreads stuff.