An error introduced in March has been fixed; Freeform tests now pass
Commit 6b005387 introduced a change where set_read_p() no longer
tested the state of the is_synthesized property. Before, if the property
was set the method would not change the state of is_read. The change
made it so the is_read propery was changed regardless of the state of
is_synthesized. This broke the 'projection functions' in the FreeForm
handler. I have reverted to the old libdap behavior, and the tests now
pass.