Fixed a bug in the DAP4 CE evaluator
The DAP4 CE evaluator was using just the stop index value
as the size of a dimension when testing if the stride value
was valid. When a dimension of length 1 (i.e., [0:0]) was
found, it's length was '0' and a stride of '1' was invalid.
I fixed this and the test now uses stop-start +1 for the
dimension length.