The change to D4Enum::compute_checksum() was correct.
This fixes the last of the issues with Enum on big-endian machines.
The problems we caused by the assumption that 64-, 32-, 16- and 8-bit
integers would align on the LSB, which is the case for little-endian
machines (but not big...). In several places where reinterpret_cast<>()
was used I first assigned the Enum _value_ to a temp variable of the
correct size and then passed the address of that to the checksum calculation
function. High cost in one sense, but really, scalar Enum variables will
be rare and Arrays of Enums are handled differently. I also have dropped
the overly-complex design that used a union. It eliminated some casts,
but was based on the erroneous assumption regarding alignment.