Modified Vector::val2buf() in support of zero-length arrays
val2buf() was throwing an InternalErr when called with a null data
pointer. I modified it so that if the pointer is null and the length()
of the vector/array is zero, it returns zero. The exception is still
thrown if the pointer is null and the length() is not zero.