Fix for Hyrax-98: Requests for Arrays with Maps fail sometimes
There are two cases where the request for an Array that contains Maps
was failing: The request didn't also name the Maps, or the request
used 'local dimension slices'. The latter case is fairly complex and
the fix here (actually both fixes here) may need revision once we 
look at how the specification should be clarified when these kinds of
CEs are used. Here is the ticket:
https://opendap.atlassian.net/browse/HYRAX-98