Fixed a problem with the TestByte, ..., classes when used for 'series values'
Some of the classes had the odd behavior that the series of values
was per-process and not per object instance. So, TestStr's value
would not repeat until a new process was started, it instead kept
increasing "... 1", "... 2", etc., across different instances of
the class. Fixing this made it easy to fix the TestD4Sequence code.