Added 'universal' tests for DMRTest
The DMR includes a checksum when it's built from received data and
that checksum is a function of word order of the sender. This made
building baselines for the DMR tests a pain. We've built a number of
tests for both big- and little-endian machines - from now on new
tests should use 'universal' baselines (built by removing the checksum
using sed).