Moved the newest tests' baselines to the 'universal' directory
It's time to switch the new DMR tests to ones that don't have to
have a separate set of baselines for bin- and little-endian
machines. It takes too long to build up the big-endian test
baselines (get a b-e VM, etc.) and we have a good set of baselines
for each arch so if issues pop up on one or the otehr arch,
it should be easy to catch them.