Fixed Hyrax-267
The DAP4 ce parser was not removing double quotes or the %<hex>
escaping, so names that needed those were not 'found' in the
dataset. Also, the grammar used by the parser had a syntax error
in it where /s in the id name were turned into dots and the
resulting names could not be found. The result was a valid names
were being rejected.