Fixed vec_resize() in Vector.
I was mixing cpacity() and resize() and that leads to some odd
behavior - I'm not sure I really understand it, but comments in
Vector::vec_resize() are a stab at an explanation. The bottom line
is that the fix is to use size() with resize() and not capacity().

I also modified TestArray so that its read() method allocates
enough space for a vector of Structures before it starts writing
to it, just to demonstrate good behavior.