DDS::get_das(): Added a test and fix for top-level 'value' attributes
It's possible to have a DMR or DDS that has top-level attributes that
are not in a container. These work in the DMR or DDS object, but not
in the DAS (which can only have containers at the top-level). The
'orphan' attributes are put in the DAP4_GLOBAL container, since these
almost certainly come from DMR objects. We might revisit this name
choice...