The DDS::print_xml and print_xml_writer now only return a '3.2' DDX
There is an issue in the BES where, for unknown reasons, a DAP2 or
DAP3.2 DDX will be returned. The DDX response should only be '3.2'.
Since there is no used for DAP2 DDX, I've removed it. This doesn't
address the underlying problem, however (why the version number in
the DDS object do not match those set by the BES's client).