Completed fixes for the DAP2 CE parser refactor.
I cannot figure out how to make UInt32 the default type for integer
constants for function arguments. I dropped back to Int32s by default.
I think this means that if functions are called with args that have
to be held in UInt32s, those function calls will not work.