Build number option (#151)
Modify the build so that the source and rpm distributions include build numbers

We've adopted a x.y.z-build version/build number scheme. Automate that so that
the build number part of the version/build number is incremented automatically
with every build of the master branch. To do this, I added a file name VERSION
that holds the number, a new option -with-build to the configure script and mods
to the .travis.yml, build-rpm.sh, and Makefile.am files.