Fixed an error in the recent fix for inverted array index values
In DAP2 we changed array slicing so that [*] means all values of
the array. In the code, we replace the * with stop value of -1 so that
syntax like [7:*] can be used to get all the values from i position
7 to the end of the dimension for a particular array. The stop == -1 was
breaking the fix I added to test if start is > stop. I hacked a change
in to special case the -1 vaule of stop. A better approach would be
to handle this in the parser to read the end value of the dimension,
but we don't have that information handy in the DAP2 parser (the DAP4
parser is much better about this sort of thing).