Dds dmr refactor (#171)
I refactored DDS and DMR to remove old code and provide better iterator access

DDS: New way of accessing the 'vars' vector makes range-based for
loops possible. 
I removed old code that was excluded with the pre processor.
Old code n configure.ac was removed.
Now DAP4 builds by default and the only form of the DDX that is supported
is the 3.2 version
Added tests for the new DDS::variables() method that returns a const & to vars.