Modify D4Group::find_var() so it can find D4Groups
When a path in a CE specifies a Group, we should be returning all the
variables in that Group. Right now libdap just throws an exception. Once
find_var() retuns the BaseType* to the Group, we should be set to mark
all of the Group's variables as part of the projection.