Added tests for the 'Group projection' feature/fix
Now projection works for Groups and nested groups both and without
member variables. That is, for normal groups (ones with child
variables) naming the group in a CE results in all the variables in
that group and its child groups being included in the response. For
Groups that only contain attributes, naming the group in the CE will
cause that group to be included in the dap (i.e., data) response. This
dubious behavior is ... well, dubious. But modulo the correctness of
attributes in the face of a constraint, it does work.

See DMRTest 243 - 249 plus some of the D4GroupTest unit tests.