Modify D4Group::find_var() so it can find D4Groups (#194)
Modify D4Group::find_var() so it can find D4Groups

When a path in a CE specifies a Group, we should be returning all the
variables in that Group. Right now libdap just throws an exception. Once
find_var() returns the BaseType* to the Group, we should be set to mark
all of the Group's variables as part of the projection.

Added print_val and print_decl methods to D4Group.
This will make it easy(ier) to add tests for the new 'all vars
in a group' feature/fix.

Added tests for the 'Group projection' feature/fix

Now projection works for Groups and nested groups both and without
member variables. That is, for normal groups (ones with child
variables) naming the group in a CE results in all the variables in
that group and its child groups being included in the response. For
Groups that only contain attributes, naming the group in the CE will
cause that group to be included in the dap (i.e., data) response. This
dubious behavior is ... well, dubious. But modulo the correctness of
attributes in the face of a constraint, it does work.

See DMRTest 243 - 249 plus some of the D4GroupTest unit tests.