D4map fix v1 (#205)
The D4Maps have two modes of creation - one when the Array is being built by, e.g.,
a handler and one when an Array is part of a deep copy operation. In the first case
it's easy to know the pointer to the Array that is the source of the Map's values.
However, during a deep copy, it's not possible to know this because the source Array
cannot be accessed since the parent Array _does not yet exist_. In this case a
second ctor is used that stores the FQN of the Map's source Array. This FQN can be
read from the Map that is being copied. When the Map is then used (the whole dataset
must exist), the FQN can be used to find the actual source Array object.

Added a ctor that supports backward compat with code in the BES

Added test for d4map-fix-v1 and verified that it fails on the master branch