The HTTPCache/HTTPConnect pair are close to working for MP access, but not quite there.
The cache will work for 'serial' access, but not full on parallel
access. If a sleep() call is added to all but the initial process,
most of the subsequnt requests see the data in the cache, but not
every subsequent process all the time.

Also, to get this, the HTTPCache code re-writes the cache index every
time new data are cached and re-reads the index before every access.
This should be made more efficient. Store a hash of the last read
version of the index file and compare it to the file on disk?