Enum dmr test fix (#279)
Fixed the DMRTest tests for enums

For test macros like PARSE and PARSE_CE, there are no word
order issues because there are no checksums. But for TRANS
and TRANS_CE, the presence of a checksum means that there
are big- and little-endian baselines. This is because the
checksum is a function of byte values but many types are
multibyte values, so the same binary data is represented
differently in the architectures and thus has a different
checksum.