Coverage 3.0 (#282)
Added comments to gcovr.config and added --sort to the coverage target.
With these changes, the Makefiles now include a target that will compute coverage
for libdap and build HTML pages that show coverage by line. Very easy to use.
At the same time, the Travis-CI 'scan' target will build a coverage.xml  file
that sonarqube can use. The two coverage measures are not exactly the same,
but they are close and each is useful in its own way.