Jhrg/xsputn part1 (#283)
This patch calls ostream::write() multiple times when data
greater than 2GB are returned as a data response. Each call
will be no more (2^31)-1 bytes. Tests with a 4 and 8GB variable
worked.

The bulk of the time (99.3%) to test these large data transfers is
in TestArray where it builds a monster array (i.e., 8GB). Looking
for ways to reduce that...

Tests pass for both --disable-threads and the default build.
Parallel I/O with libdap now works with variables that are 2^33 bytes.

Added a new class MarshallerFuture (header only) and unit tests.
This is not used, but could be a replacement for MarshallerThread,
which uses the old and cumbersome pthreads library. This new class
uses the C++11/14 futures and also supports shared_ptr buffers for
potential reductions in memory use.