Corrected 3 baselines; all tests pass.
There are three tests that all are simply an empply DAS
the return an error, but return nothing to stdout. The
autotest code diffs the output of stdout against the
basline (which is empty on the master branch now) and
that will pass. This code is more restrictive and diffs
the output on stderr. I updated the baselines to use the
stderr output. This should work on the master branch as well.