Fixed null baselines. See HYRAX-1839. (#292)
A bunch of our libdap4 regression tests had empty baseline files. I fixed this, 
and it was not as widespread as it could have been. But the bug was that our
test macro for the 'expr-test' program was using a logical OR to test if the 
baseline was equal to the stdout OR stderr output. If the expr-test program 
ran correctly, then the stderr output was null, so the tests will null baselines 
passed no matter what expr-test put on stdout!