jhrg/libdap4-305, HYRAX-1948, serialization bug (#306)
The DAP4 specification was/is confusing about the order of serialization
of datasets that contain Groups. The order is specified, but in the best 
form of a specification, the explanation is spread across two parts of the
spec.

The variables in the root group are now serialized first, then variables in
child groups.

Added tests for this bug and a fix for the broken implementation in libdap.

Tested the fix against the NSIDC OA application, which is the only NGAP
we are aware of that makes direct use of the DAP4 binary responses using
client code from Unidata.