The 'scanner test' option in dmr-test now works.
What is not so good is that a problem shows up. Using the new
syntax [(45.6)] fails because the '.' scans as a PATH_SEP, and
that breaks everything. This is why I added NUMBER before. But,
that causes its own set of parsing issues. Try FLOAT and add
rules in the parser for both [(FLOAT)] and [(WORD)]?